+++ components/31-molecules/fields/checkbox-with-filter/checkbox_filter.functions.js
@@ -96,22 +96,6 @@
     );
 
     /**
-     * Container to display the placeholder result text prior to filtering.
-     * @type {Element}
-     */
-    const resultSpanPlaceholder = elem.querySelector(
-      options.resultSpanPlaceholder || '.checkbox-filter__result-placeholder'
-    );
-
-    /**
-     * Container wrapping the resultspan.
-     * @type {Element}
-     */
-    const resultSpanWrapper = elem.querySelector(
-      options.resultSpanWrapper || '.checkbox-filter__result-wrapper'
-    );
-
-    /**
      * Container wrapping the countspan.
      * @type {Element}
      */
@@ -268,24 +252,6 @@
     const updateResult = (resultCount) => {
       if (resultSpan) {
         resultSpan.textContent = resultCount;
-      }
-
-      if (resultSpanWrapper) {
-        if (filterfield.value === '') {
-          resultSpanWrapper.classList.add('hidden');
-        }
-        else {
-          resultSpanWrapper.classList.remove('hidden');
-        }
-      }
-
-      if (resultSpanPlaceholder) {
-        if (filterfield.value === '') {
-          resultSpanPlaceholder.classList.remove('hidden');
-        }
-        else {
-          resultSpanPlaceholder.classList.add('hidden');
-        }
       }
     };